requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
requests==2.31.0
python-dotenv==1.0.0
python-nmap==0.7.4
aiohttp==3.8.5
pydantic==1.10.11
